requirements.txt
streamlit==0.75.0
pandas==1.2.0
numpy==1.19.5
Pillow==8.1.0
